\name{MouseArraymanifest}
\alias{Mouse Arraymanifest}
\docType{data}
\title{
  Manifest object for the Mouse array
}
\description{
  This object contains the array design for the mouse array.
}
\usage{data(MouseArraymanifest)}
\details{

}
\format{
  An object of class \code{MouseArraymanifest}
}
\references{
  None.
}
\seealso{
  The description of the class in the minfi package,
}
\examples{
data(MouseArraymanifest)
}
\keyword{datasets}
